node.c: Iterate over attributes in attributes array offsets order.
When not printing debug information we now iterate over the attributes
in the correct order as specified in the offsets array. This fixes an
issue where attributes were returned out of order if they were packed
out of order in the long entry (they would be returned by the order of
appearance in the long entry).

When printing debug information we now print the proper attribute index
of each attribute instead of incrementing the index for every
encountered attribute.