node.c: Fix up leaf data prinouts for the 0x21 bitmap tree.
We still don't know what all the fields mean but it seems pretty clear
where the bitmap data is, and the length of it... and the first number
appears to be the block number. We still haven't run any validation on
the bitmap data but that will be the next step.